requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
questionary==2.0.1
rich==13.7.1
loguru==0.7.2
